Refactor computation code
