+++ packages/web_benchmarks/lib/client.dart
@@ -11,8 +11,10 @@ import 'package:web/web.dart';
 
 import 'src/common.dart';
 import 'src/recorder.dart';
+import 'src/computations.dart';
 
 export 'src/recorder.dart';
+export 'src/computations.dart';
 
 /// Signature for a function that creates a [Recorder].
 typedef RecorderFactory = Recorder Function();
